Fixing verification assessment in plot_heatmap #101 (#102)
* Fixing verification assessment in plot_heatmap

Removed the verification check for the columns finalX and finalY from the plot_heatmap function.

* Fixing verification assessment in plot_heatmap

Removed the verification check for the columns finalX and finalY from the plot_heatmap function. Also updated the test functions correspondingly and used roxygen to update the .Rd file for plot_heatmap. The changes have been reflected in the website documentation as well.

* Increment version number

* Increment version number and update the yaml files

Fixes the pull request issues in the fix for #101

* Incrementing version to 0.2.0.9000

Corrected the version increment to 0.2.0.9000 instead of making it
0.3.0.x.

closes #101